explicit index type

+++ src/rhetoric/file_stream.h
@@ -1,13 +1,15 @@
 #pragma once
 
+#include "./std_dependency.h"
+#include "./unix_dependency.h"
+
 #include "./crt.h"
 #include "./file_path.h"
+#include "./format_macro.h"
 #include "./posix_error.h"
 #include "./ptr.h"
 #include "./result.h"
-#include "./std_dependency.h"
 #include "./try_macro.h"
-#include "./unix_dependency.h"
 
 namespace rhetoric {
     class FileStream : public std::enable_shared_from_this<FileStream> {
@@ -25,16 +27,16 @@ namespace rhetoric {
         Result<Ptr<Data>> Read();
 
         RHETORIC_NO_DISCARD
-        Result<Ptr<Data>> Read(int size);
+        Result<Ptr<Data>> Read(size_t size);
 
         RHETORIC_NO_DISCARD
-        Result<int> ReadToBytes(void * bytes, int size);
+        Result<size_t> ReadToBytes(void * bytes, size_t size);
 
         RHETORIC_NO_DISCARD
         Result<None> Write(const Ptr<const Data> & data);
 
         RHETORIC_NO_DISCARD
-        Result<None> WriteFromBytes(const void * bytes, int size);
+        Result<None> WriteFromBytes(const void * bytes, size_t size);
 
         RHETORIC_NO_DISCARD
         Result<None> Seek(int64_t offset, int whence);
